Use http_host not proxy_port for setting Host header
This is imperfect, but only really breaks for badly behaving clients
that get what they deserve. proxy_port is not what we want. We want to
allow the upstream service to create URLs to return in response bodies
that match what the client used to contact the service. If we set
anything hard-coded here we will be wrong if there are extra layers of
forwarding/proxying. So just pass along what came in on the request
and hope for the best.