Align nginx search endpoint regexes with prod
NOTE: These are BAD regexes, because the ones in prod are also bad!  In prod, we aren't actually going to Erchef for the search endpoint, because the search endpoint regex (the first one in this patch) is not specific enough, allowing the second regex (which dispatches everything to Ruby Chef) to override it (note that "search" is still present in that regex!).  Note that nginx applies the *most specific* location directive to a given request.

I'm changing this so PIAB will behave the same as our production platform BUT THESE ARE WRONG AND NEED TO BE FIXED.